pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
    "setuptools-rust",
    "breezy>=3.3.1",
    "python_debian",
    "debmutate",
]
build-backend = "setuptools.build_meta"

[tool.mypy]
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "psycopg2.*",
    "distro_info.*",
    "decopy.*",
    "gpg.*",
    "iniparse.*",
    "Levenshtein.*",
    "semver.*",
]
ignore_missing_imports = true

[project]
name = "lintian-brush"
authors = [{name = "Jelmer Vernooĳ", email = "jelmer@debian.org"}]
description = "Automatic lintian issue fixer"
requires-python = ">=3.7"
dependencies = [
    "breezy>=3.3.1",
    "python_debian",
    "debmutate[watch,debcargo]",
    "distro-info",
    "upstream-ontologist[readme,pyproject,setup.cfg,debian_rules,debian_watch,debian_changelog,cargo]>=0.1.29",
    "ruamel.yaml",
    "psycopg2",
    "semver",
    "tomlkit",
    "python_Levenshtein",
    "iso8601",
    "iniparse",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://salsa.debian.org/jelmer/lintian-brush"
Repository = "https://salsa.debian.org/jelmer/lintian-brush"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["py"]
include = ["lintian_brush"]

[tool.setuptools.dynamic]
version = {attr = "lintian_brush.__version__"}

[tool.ruff]
target-version = "py37"
line-length = 79

[tool.ruff.lint]
select = [
    "ANN",
    "D",
    "E",
    "F",
    "I",
    "UP",
]
ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN206",
    "ANN401",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "D300",
    "D417",
    "E501",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["lintian_brush"]
